Move ProductBox to features
